common: use libsoc original function to get the alias
Instead of using the functions we added to libsoc to get the I2C, PWM, and SPI
alias, use 'conffile_get_int()' from libsoc. Although it is interesting to do
a pull request to add this wrappers to the official libsoc repository, this way
we do not depend on when this happens.

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>