gpio: added libgpiod support
1)
From the user point of view the main change is the addition of function
ldx_gpio_request_by_controller() which selects the gpio to use using following
arguments:
   - controller: The controller name or alias of the GPIO to request.
   - line: The line number of the GPIO to request.

Once this function is called, the rest of the changes are transparent for the user
but the ligdigiapix will use underlying libgpiod calls instead of legacy sysfs calls.

2)
The configuration file entries change from:
USER_LED = 505
to:
USER_LED = mca-gpio,12

3)
Current porting of libgpiod has following TODOs:
   - Check request_mode:
     The sysfs implementation just forwarded this argument to the libsoc library
     with options REQUEST_SHARED, REQUEST_GREEDY and REQUEST_WEAK. For libgpiod
     it has to be implemented manually and all options may not be available.
   - Implement set_debounce:
     May not be supported in gpiod. Didn't see anything available.
     Could we go through the sysfs?

4)
Current implementation keeps all the sysfs functionality as it was originally.
In most of the calls, the library checks gpio->kernel_number and if it's '-1' it
makes use of new libgpiod calls; if different, use legacy sysfs calls.
It's likely that we want to implement following change:
   - Remove legacy sysfs implementation and after calling legacy ldx_gpio_request()
     or ldx_gpio_request_by_alias() entry functions, get corresponding controller
     label and line and commute to libgpiod calls.
     Initial attempt of implementing this was not straightforward:
        - Neither the character devices not the libgpiod library provide information
          of the 'base' number of a controller that could be used to match the
          legacy function 'kernel_number' parameter to a controller/line.
        - To achive the matching, the /sys/class/gpio directory has to be listed
          and, for each entry, the 'base' and 'ngpio' files parsed to select the
          desired gpiochip and then make use of it 'label' entry.

https://jira.digi.com/browse/DEL-6425

Signed-off-by: Hector Bujanda <hector.bujanda@digi.com>