ccimx: fix a fail when parsing the output from fgets
Since the migration to dey-3.0, there was a different behavior
in the fgets function in line 284 of the common.c.

The string returned by the command have some NULL characters
and it is not well parsed by the code, giving some failures and
mismatches.

With this change we only get the final part that it is used
to determine the platform without NULL characters keeping
backward compatibility.

https://onedigi.atlassian.net/browse/DEL-7544

Signed-off-by: Francisco Gil Martinez <francisco.gilmartinez@digi.com>