i2c: rework function to list i2c buses
The original function ran from 0 to MAX_I2C_BUSES trying to
find buses listed as /dev/i2c-<n>. The problem with this is
that not all buses are listed with an ordered index, and the
constant was limiting the amount of buses it could find.

Rework the function to list 'any' available index using a shell
command and parsing the actual available ports.

While on it, increase the buses limit from 5 to 10.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-7480