bindings: python: first implementation of Python bindings
This commit adds the foundations of the Python bindings for the library. These
bindings are installed in the system in the form of a new Python module named
"digi.apix". Users can call and execute the C library methods from Python
applications using the object oriented API exposed in the new module.

The first implemented interface is the GPIO.

Signed-off-by: David Escalona <david.escalona@digi.com>