can: fix control mode validation
This commit fixes the mechanism to validate if a modification operation for the
control mode was successful. Users can use a mask to select modes they want to
control and flags to determine if they want to turn the selected modes on or
off, we only must compare the flags under the mask control.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>