network: mechanism for mapping interface names to NM devices and connections
This network API is used in the Cloud Connector to get the "main" network
interface information (the interface used to connect to DRM server)
The process:
 1. Resolves the IP of the configured DRM end point
 2. Connects a socket to this IP and the corresponding port (TCP or SSL port)
 3. Gets the address to which this socket is bound
 4. Loops over the available interfaces to compare the bound address with the
    interface IP. The one that matches is the "main" network interface.

The problem is when the interface name does not match the name that Network
Manager uses for the corresponding device nor for the connection.
For example, using a cellular modem the mapped interface name does not match
with NM device and connection names. For a XBee 3 LTE Cat1, the interface name
is 'wwan0', but NM uses 'ttyACM1' as device name and 'cellular' for the
connection name.

This commit implements a way of mapping the interface name to the NM device and
connection names, to use these lasts ones when NM is used to interact with any
network interface.

Tested with a XBee 3 LTE Cat1.

https://onedigi.atlassian.net/browse/DEL-8459

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>