can: add thread condition signal to sync start of worker thread
This commit adds a thread signal that is used to start the
working thread when the application is ready. This
is necessary because in some occasions the thread started
before the RX Handler and locked the mutex, which could
result in delays during the start of the apix_can_recv_example.

https://onedigi.atlassian.net/browse/DEL-8833

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>
Signed-off-by: Mike Engel <Mike.Engel@digi.com>