cccs: python: remove logging from CCCS APIx and rely on exceptions
Syslog is not available as a Python package in the DEY distribution, and the DEY
Python APIx does not use any logging mechanism. Instead, all errors are reported
via exceptions.

For this reason, remove the logging entries from the CCCS APIx and rely solely on
thrown exceptions for error reporting.

https://onedigi.atlassian.net/browse/OEMIOTD-235

Signed-off-by: David Escalona <david.escalona@digi.com>